refactor org.apache.shardingsphere.encrypt.yaml package
