+++ shardingsphere-features/shardingsphere-encrypt/shardingsphere-encrypt-common/src/main/java/org/apache/shardingsphere/encrypt/yaml/config/rule/YamlEncryptColumnRuleConfiguration.java
@@ -15,24 +15,26 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.encrypt.yaml.config;
+package org.apache.shardingsphere.encrypt.yaml.config.rule;
 
 import lombok.Getter;
 import lombok.Setter;
 import org.apache.shardingsphere.infra.yaml.config.YamlConfiguration;
 
-import java.util.Properties;
-
 /**
- * Encrypt strategy configuration for YAML.
+ * Encrypt column rule configuration for YAML.
  */
 @Getter
 @Setter
-public final class YamlEncryptStrategyConfiguration implements YamlConfiguration {
+public final class YamlEncryptColumnRuleConfiguration implements YamlConfiguration {
     
-    private String name;
+    private String logicColumn;
     
-    private String type;
+    private String cipherColumn;
     
-    private Properties props = new Properties();
+    private String assistedQueryColumn;
+    
+    private String plainColumn;
+    
+    private String encryptorName;
 }
